Rename window files
