+++ Sprocket/Core/game_loop.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Sprocket/Core/events.h>
-#include <Sprocket/Core/Window.h>
+#include <Sprocket/Core/window.h>
 #include <Sprocket/Core/log.h>
 #include <Sprocket/Core/timer.h>
